RavenDB-19659 - fix failing test
